requirements.txt
# =============================================================================
# Backend Dependencies — Klasifikasi Tanaman Obat
# =============================================================================
# Install semua: pip install -r requirements.txt
# =============================================================================

# Machine Learning & Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Science & Visualisasi
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Image Processing
Pillow>=10.0.0

# Web API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Jupyter Notebook (untuk VS Code)
jupyter>=1.0.0
ipykernel>=6.0.0
